+++ app/components/Input/index.tsx
@@ -1,4 +1,6 @@
 'use client';
+import { createContext, useState } from 'react';
+
 import {
   CheckInputProps,
   DateInputProps,
@@ -9,10 +11,9 @@ import {
   TextAreaInputProps,
   TextInputProps,
 } from './type';
-import IconError from '@/assets/images/icon-error.svg';
 import { Body, Caption } from '../../constants/Typography/TypographyList';
-import { createContext, useState } from 'react';
 import useInputContext from '@/hooks/useInputContext';
+import IconError from '@/assets/images/icon-error.svg';
 
 export const defaultInputContext: InputContextProps = {
   isValid: false,
