+++ app/components/Schedule/AddScheduleModal/PickCalendar.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import CalendarForm from '../../CalendarForm';
-import { handleformattedDate } from '@/components/Account/PetInfoForm';
+
 import Input from '@/components/Input';
 import { SCHEDULE_TYPE } from '../constants';
+import Calendar from '../../Calendar/CalendarPicker';
 
 export type DateType = 'startTime' | 'endTime';
 
@@ -16,6 +16,7 @@ interface PickCalendarProps {
     variant: 'hh' | 'mm',
   ) => void;
 }
+
 const PickCalendar = ({
   scheduleTime,
   type,
@@ -25,11 +26,13 @@ const PickCalendar = ({
   return (
     <li className='bg-grayColor-10'>
       <Input name={SCHEDULE_TYPE.ADDRESS}>
-        <CalendarForm
-          initDate={new Date(scheduleTime.date)}
-          handleDayClick={(day: Date) => handleChangeDate(day, type)}
-          headerType='center'
-        />
+        <Calendar>
+          <Calendar.YYYYMMPicker type='center' />
+          <Calendar.Date
+            handleClickDay={(day: Date) => handleChangeDate(day, type)}
+          />
+        </Calendar>
+
         <div className='px-3 pb-5'>
           <Input.Label>시간 설정</Input.Label>
           <div className='bg-white group-focus:border-extra-active flex justify-center border border-extra-border rounded-md overflow-hidden px-[6px] py-4'>
