+++ app/components/Schedule/AddScheduleModal/TimeFormatter.tsx
