+++ app/components/Schedule/type.ts
